kiss

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: jetownle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/08 11:44:49 by jetownle          #+#    #+#             */
-/*   Updated: 2019/05/15 16:56:57 by jetownle         ###   ########.fr       */
+/*   Updated: 2019/05/21 01:58:29 by jetownle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (((unsigned char *)s1)[i] == ((unsigned char *)s2)[i] && (i < n))
